+++ src/index.test.js
@@ -1,3 +1,7 @@
 export * from 'angular-mocks';
 export * from './checkers/index.test';
 export * from './scopeTypes.test';
+import chai from 'chai';
+import sinonChai from 'sinon-chai';
+
+chai.use(sinonChai);