+++ src/scopeTypes.js
@@ -2,10 +2,19 @@ import angular from 'angular-fix';
 import apiCheckFactory from 'api-check';
 import checkerFactories from './checkers';
 
+const defaultOutput = {prefix: 'api-check-angular'};
 
-export default function scopeTypesFactory({disabled = false} = {disabled: false}) {
+export default scopeTypesFactory;
+
+function scopeTypesFactory({
+  disabled = false,
+  output = defaultOutput
+  } = {
+  disabled: false,
+  output: defaultOutput
+}) {
   const scopeTypes = apiCheckFactory({
-    output: {prefix: 'api-check-angular'},
+    output: output,
     disabled
   });
   // manually adding checkers so we have an instance of scopeTypes to pass them.
@@ -18,7 +27,10 @@ export default function scopeTypesFactory({disabled = false} = {disabled: false}
   return scopeTypes;
 
   function validateDirective(ddo) {
-    scopeTypes.warn(scopeTypes.ddo, ddo, {prefix: 'creating directive'});
+    if (scopeTypes.config.disabled) {
+      return ddo;
+    }
+    scopeTypes.warn(scopeTypes.ddo, ddo, {prefix: 'creating directive with scopeTypes'});
     // Would really love to not have to extend the ddo's controller
     // like this. Suggestions welcome!
     ddo.controller = extendController(ddo.controller, scopeTypesController);
@@ -29,14 +41,18 @@ export default function scopeTypesFactory({disabled = false} = {disabled: false}
       if (ddo.bindToController) {
         context = context[ddo.controllerAs];
       }
-      let typeDefinition = ddo.scopeTypes;
-      if (angular.isFunction(typeDefinition)) {
-        typeDefinition = typeDefinition(scopeTypes);
-      }
-      angular.forEach(typeDefinition, function(check, name) {
+
+      let typeDefinitions = ddo.scopeTypes(scopeTypes);
+      scopeTypes.warn(
+        scopeTypes.objectOf(scopeTypes.func), typeDefinitions, {prefix: `getting scope types for ${ddo.name}`}
+      );
+
+      $scope.$scopeTypesResults = {__passed: 0, __failed: 0};
+
+      angular.forEach(typeDefinitions, (check, name) => {
         if (!angular.isDefined(context[name])) {
           let prefix = ddo.controllerAs ? ddo.controllerAs + '.' : '';
-          const stopWatching = $scope.$watch(`${prefix}${name}`, function(value, oldValue) {
+          const stopWatching = $scope.$watch(`${prefix}${name}`, (value, oldValue) => {
             if (value !== oldValue) {
               stopWatching();
               checkOption(check, name);
@@ -49,7 +65,27 @@ export default function scopeTypesFactory({disabled = false} = {disabled: false}
 
 
       function checkOption(checker, name) {
-        scopeTypes.warn(checker, context[name], {prefix: `${ddo.name}Directive`});
+        $scope.$scopeTypesResults[name] = scopeTypes.warn(
+          checker, context[name], {prefix: `${ddo.name}Directive for "${name}"`}
+        );
+        updateData();
+      }
+
+      function updateData() {
+        let passedCount = 0;
+        let failedCount = 0;
+        const ignore = ['__passed', '__failed'];
+        angular.forEach($scope.$scopeTypesResults, (result, name) => {
+          if (ignore.indexOf(name) === -1) {
+            if (result.passed) {
+              passedCount++;
+            } else {
+              failedCount++;
+            }
+          }
+        });
+        $scope.$scopeTypesResults.__passed = passedCount;
+        $scope.$scopeTypesResults.__failed = failedCount;
       }
 
     }