+++ CHANGELOG.md
@@ -1,2 +1,14 @@
+# 2.0.0
+
+## Breaking changes
+
+- `scopeTypes` must be a function that accepts the api-check instance and returns the api (perf improvement)
+- Changing what is exported to be the `scopeTypesFactory` function.
+
+## New Features
+
+- Adding `disabled` option.
+- Improved the API, adding more tests, adding example.
+
 # 1.0.0